Add Regal linter config, and fix some reported issues
Kicking the tires of my new open source linter for Rego,
[Regal](https://github.com/styrainc/regal), and wanted to test it
against a real-world policy library. This was a great exercise, as
it uncovered a few quirks in Regal, which have now been fixed.

I figured that I'd share the results with you in a PR, so here we
are.

This PR fixes a number of issues reported:

- [redundant-alias](https://github.com/StyraInc/regal/blob/main/docs/rules/imports/redundant-alias.md)
- [redundant-data-import](https://github.com/StyraInc/regal/blob/main/docs/rules/imports/redundant-data-import.md)
- [use-in-operator](https://github.com/StyraInc/regal/blob/main/docs/rules/style/use-in-operator.md)
- [rule-shadows-builtin](https://github.com/StyraInc/regal/blob/main/docs/rules/bugs/rule-shadows-builtin.md)
- [custom-in-construct](https://github.com/StyraInc/regal/blob/main/docs/rules/idiomatic/custom-in-construct.md)

Provided is also a configuration for Regal that disables a number of rules
mostly related to style, which someone might want to look into one day, but
they are not critical.

If you like it, perhaps we could have it integrated into the CI pipeline to have
the linter run on PRs, etc.

Finally, I added support for `print` statements in the test runner, as that greatly
simplifies debugging the tests.

Signed-off-by: Anders Eknert <anders@styra.com>