feat: Add unauthenticated service control
This commit adds the unauthenticated service control to the `C-0256-exposuretointernet.json` and `C-0266-exposuretointernet-gateway.json` files. The control is categorized under "Initial Access" and is part of the "workload-unauthenticated-service" attack track.

Note: This commit message follows the established convention of starting with a type prefix (feat for feature) and providing a concise and descriptive summary of the changes made.
Signed-off-by: kooomix <eranm@armosec.io>