chore: tidy up use of spec_template_spec_patterns
The

    spec_template_spec_patterns := {"Deployment","ReplicaSet","DaemonSet","StatefulSet","Job"}
    spec_template_spec_patterns[wl.kind]

is common pattern in multiple rules. This commit fixes instances, where
the second line of this pattern was missing. Effectively rendering the
first line of the pattern to be useless assignment.

For future reference, I found, it might be useful to complete this
pattern properly, even though this commit does not have any logical
effect on the scanner's RuleResults.

Found only 3 occurrences of this issue.

Signed-off-by: Šimon Lukašík <isimluk@fedoraproject.org>